fix(#121): support nested keys
Nested keys were supported to some degree as in the model would update
but it would not properly set the form control.

This was due the fact that the formgroup would not nest when a
fieldgroup was specified.

This was fixed by making registerFormControls context sensitive in the
sense that the nested model and nested formgroup will be passed if in
the formfieldconfig the key was specified.

Updated the demo to showcase that indeed the value will be set at the
start.