chore(schematics): fix schematics pkg version
Schematics has a different build process from the other `ngx-formly`
packages and therefore needs a slightly modified build process.
To not alter too much, add a step to overwrite the dist
`package.json` to have the "FORMLY_VERSION" key that is used
to update to correct Formly version on builds.