feat(core): support functions returning Observables from expressions
Expressions keys that end with $ are treated as returning an Observable value, if the value is a
function it is executed and expected to return an Observable. This allows for accessing the field
data via the parameter in the same manor as typical expression, as well as more complex situations
that may require multiple other observables conditionaly to be resolved.

fix #1818, fix #2013